working on the func of wgs
