--- depmapomics/rna.py
@@ -1,33 +1,0 @@
-#rna.py
-import pandas as pd
-
-def addSamplesRSEMToMain(input_filenames, main_filename):
-  """
-  given a tsv RNA files from RSEM algorithm, merge it to a tsv set of RNA data
-
-  Args:
-  ----
-    input_filenames: a list of dict like file path in Terra gs://, outputs from the rsem pipeline
-    main_filename: a dict like file paths in Terra gs://, outputs from rsem aggregate
-  """
-  genes_count = pd.read_csv('temp/' + main_filename['rsem_genes_expected_count'].split('/')[-1],
-                            sep='\t', compression='gzip')
-  transcripts_tpm = pd.read_csv('temp/' + main_filename['rsem_transcripts_tpm'].split('/')[-1],
-                                sep='\t', compression='gzip')
-  genes_tpm = pd.read_csv('temp/' + main_filename['rsem_genes_tpm'].split('/')[-1],
-                          sep='\t', compression='gzip')
-
-  for input_filename in input_filenames:
-    name = input_filename['rsem_genes'].split('/')[-1].split('.')[0].split('_')[-1]
-    rsem_genes = pd.read_csv('temp/' + input_filename['rsem_genes'].split('/')[-1], sep='\t')
-    rsem_transcripts = pd.read_csv('temp/' + input_filename['rsem_isoforms'].split('/')[-1], sep='\t')
-    genes_count[name] = pd.Series(rsem_genes['expected_count'], index=rsem_genes.index)
-    transcripts_tpm[name] = pd.Series(rsem_transcripts['TPM'], index=rsem_transcripts.index)
-    genes_tpm[name] = pd.Series(rsem_genes['TPM'], index=rsem_genes.index)
-
-  genes_count.to_csv('temp/' + main_filename['rsem_genes_expected_count'].split('/')[-1], sep='\t',
-                     index=False, index_label=False, compression='gzip')
-  transcripts_tpm.to_csv('temp/' + main_filename['rsem_transcripts_tpm'].split('/')[-1], sep='\t',
-                         index=False, index_label=False, compression='gzip')
-  genes_tpm.to_csv('temp/' + main_filename['rsem_genes_tpm'].split('/')[-1], sep='\t',
-                   index=False, index_label=False, compression='gzip')
